source/cdk/requirements.txt
boto3
aws-cdk-lib==2.51.1
constructs>=10.0.0,<11.0.0
